Corrige detección de input y cerrado de ventana
Finalmente pude corregir las fallas del sistema del teclado, y ahora
se pueden enviar señales al apretar teclas. Era este también el
problema que ocasionaba que no pudiera cerrarse la ventana: se esperaba
que se detectaran eventos SDL constantemente, y al no hacerlo, la
ventana no funcionaba correctamente.

Ahora existe otro problema: aunque funciona bien a grandes rasgos, la
pantalla suele mostrar algunos desperfectos, con pixeles que se quedan
pegados entre otras cosas.

Además, al compilar el programa en modo release el programa crashea
instantáneamente, supuestamente por un segfault, por lo que también
será necesario solucionar eso.

+++ include/CHIP-8/io.h
@@ -23,11 +23,11 @@ public:
      IO (String title, Pixels width, Pixels height, Scale scale);
     ~IO ();
 
-    void    clear      ();
-    void    draw       (const Sprite& sprite, Coord x, Coord y);
-    uint8_t last_key   ();
-    uint8_t wait_key   ();
-    void    update     ();
+    void     clear    ();
+    void     draw     (const Sprite& sprite, Coord x, Coord y);
+    uint16_t last_key ();
+    uint8_t  wait_key ();
+    void     update   ();
 
 private:
     struct RGBA {
